Fix default-namespace XML parsing.
While we're at it, add a parsing test that actually checks that parsing
occurred.

Nokogiri knows what namespace every element is in, regardless of whether
it's the default xmlns or one with an alias. If you XPath against it,
you have to give it an XPath that specifies the right namespace - though
there's no valid prefix in the default-namespace case. The prefix
doesn't have to match the prefix being used in the document; it just
needs to map to a key in the namespace map you give it at XPath time.
Since we already have the default namespace in our map as the 'xmlns'
key, we fix up the XPath in this case to require the element XPath to
use the 'xmlns' prefix.